requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
openpyxl==3.1.5
pandas==2.2.2
